refactor(procedures): implement class-based architecture for DALs
