+++ src/orpc/lease/index.ts
@@ -1,11 +1,4 @@
-import {
-  acquire as acquireLease,
-  requestTakeover as createTakeoverRequest,
-  heartbeat as heartbeatLease,
-  listForTournament,
-  release as releaseLease,
-  respondTakeover as respondTakeoverRequest,
-} from './lease.dal';
+import { LeaseDAL } from './dal';
 import {
   AcquireLeaseSchema,
   HeartbeatLeaseSchema,
@@ -13,56 +6,62 @@ import {
   ReleaseLeaseSchema,
   RequestTakeoverSchema,
   RespondTakeoverSchema,
-} from './lease.dto';
+} from './dto';
 import { authedProcedure } from '@/orpc/middleware';
 
 export const acquire = authedProcedure
   .input(AcquireLeaseSchema)
   .handler(async ({ context, input }) => {
-    return acquireLease({
+    const lease = await LeaseDAL.acquire({
       ...input,
       adminId: context.user.id,
     });
+    return lease;
   });
 
 export const heartbeat = authedProcedure
   .input(HeartbeatLeaseSchema)
   .handler(async ({ context, input }) => {
-    return heartbeatLease({
+    const lease = await LeaseDAL.heartbeat({
       ...input,
       adminId: context.user.id,
     });
+    return lease;
   });
 
 export const release = authedProcedure
   .input(ReleaseLeaseSchema)
   .handler(async ({ context, input }) => {
-    return releaseLease({
+    const lease = await LeaseDAL.release({
       ...input,
       adminId: context.user.id,
     });
+    return lease;
   });
 
 export const requestTakeover = authedProcedure
   .input(RequestTakeoverSchema)
   .handler(async ({ context, input }) => {
-    return createTakeoverRequest({
+    const request = await LeaseDAL.requestTakeover({
       ...input,
       adminId: context.user.id,
     });
+    return request;
   });
 
 export const respondTakeover = authedProcedure
   .input(RespondTakeoverSchema)
   .handler(async ({ context, input }) => {
-    return respondTakeoverRequest({
+    const response = await LeaseDAL.respondTakeover({
       ...input,
       adminId: context.user.id,
     });
+    return response;
   });
 
 export const listForTournamentLeases = authedProcedure
   .input(ListLeasesForTournamentSchema)
   .handler(async ({ input }) => {
-    return listForTournament(input);
+    const leases = await LeaseDAL.listForTournament(input);
+    return leases;
   });
