+++ src/orpc/matches/index.ts
@@ -12,25 +12,8 @@ import {
   SwapSlotsSchema,
   UpdateMatchSchema,
   UpdateScoreSchema,
-} from './matches.dto';
-import {
-  assignSlot,
-  create,
-  deleteMatch,
-  findByGroupId,
-  findById,
-  findByTournamentId,
-  generateBracket,
-  regenerateBracket,
-  resetBracket,
-  setLock,
-  setWinner,
-  shuffleBracket,
-  swapParticipants,
-  swapSlots,
-  update,
-  updateScore,
-} from './matches.dal';
+} from './dto';
+import { MatchDAL } from './dal';
 import { authedProcedure } from '@/orpc/middleware';
 
 export const listMatches = authedProcedure
@@ -42,10 +25,12 @@ export const listMatches = authedProcedure
   )
   .handler(async ({ input }) => {
     if (input.groupId) {
-      return findByGroupId(input.groupId);
+      const matches = await MatchDAL.findByGroupId(input.groupId);
+      return matches;
     }
     if (input.tournamentId) {
-      return findByTournamentId(input.tournamentId);
+      const matches = await MatchDAL.findByTournamentId(input.tournamentId);
+      return matches;
     }
     throw new Error('Either groupId or tournamentId is required');
   });
@@ -53,86 +38,110 @@ export const listMatches = authedProcedure
 export const getMatch = authedProcedure
   .input(z.object({ id: z.string() }))
   .handler(async ({ input }) => {
-    const match = await findById(input.id);
-    if (!match) throw new Error('Match not found');
+    const match = await MatchDAL.findById(input.id);
+    if (!match) {
+      throw new Error('Match not found');
+    }
     return match;
   });
 
 export const createMatch = authedProcedure
   .input(CreateMatchSchema)
   .handler(async ({ input }) => {
-    return create(input);
+    const match = await MatchDAL.create(input);
+    return match;
   });
 
 export const updateMatch = authedProcedure
   .input(UpdateMatchSchema)
   .handler(async ({ input }) => {
     const { id, ...data } = input;
-    return update(id, data);
+    const match = await MatchDAL.update(id, data);
+    return match;
   });
 
 export const removeMatch = authedProcedure
   .input(z.object({ id: z.string() }))
   .handler(async ({ input }) => {
-    return deleteMatch(input.id);
+    const match = await MatchDAL.deleteMatch(input.id);
+    return match;
   });
 
 export const generateBracketEndpoint = authedProcedure
   .input(GenerateBracketSchema)
   .handler(async ({ input, context }) => {
-    return generateBracket(input, context.user.id);
+    const matches = await MatchDAL.generateBracket(input, context.user.id);
+    return matches;
   });
 
 export const shuffleBracketEndpoint = authedProcedure
   .input(ShuffleBracketSchema)
   .handler(async ({ input, context }) => {
-    return shuffleBracket(input.groupId, context.user.id);
+    const matches = await MatchDAL.shuffleBracket(
+      input.groupId,
+      context.user.id
+    );
+    return matches;
   });
 
 export const regenerateBracketEndpoint = authedProcedure
   .input(RegenerateBracketSchema)
   .handler(async ({ input, context }) => {
-    return regenerateBracket(input.groupId, context.user.id);
+    const matches = await MatchDAL.regenerateBracket(
+      input.groupId,
+      context.user.id
+    );
+    return matches;
   });
 
 export const resetBracketEndpoint = authedProcedure
   .input(ResetBracketSchema)
   .handler(async ({ input, context }) => {
-    return resetBracket(input.groupId, context.user.id);
+    const matches = await MatchDAL.resetBracket(input.groupId, context.user.id);
+    return matches;
   });
 
 export const setLockEndpoint = authedProcedure
   .input(SetLockSchema)
   .handler(async ({ input }) => {
-    return setLock(input);
+    const match = await MatchDAL.setLock(input);
+    return match;
   });
 
 export const updateScoreEndpoint = authedProcedure
   .input(UpdateScoreSchema)
   .handler(async ({ input, context }) => {
-    return updateScore(input, context.user.id);
+    const score = await MatchDAL.updateScore(input, context.user.id);
+    return score;
   });
 
 export const setWinnerEndpoint = authedProcedure
   .input(SetWinnerSchema)
   .handler(async ({ input, context }) => {
-    return setWinner(input, context.user.id);
+    const winner = await MatchDAL.setWinner(input, context.user.id);
+    return winner;
   });
 
 export const swapParticipantsEndpoint = authedProcedure
   .input(SwapParticipantsSchema)
   .handler(async ({ input, context }) => {
-    return swapParticipants(input, context.user.id);
+    const participants = await MatchDAL.swapParticipants(
+      input,
+      context.user.id
+    );
+    return participants;
   });
 
 export const assignSlotEndpoint = authedProcedure
   .input(AssignSlotSchema)
   .handler(async ({ input, context }) => {
-    return assignSlot(input, context.user.id);
+    const slot = await MatchDAL.assignSlot(input, context.user.id);
+    return slot;
   });
 
 export const swapSlotsEndpoint = authedProcedure
   .input(SwapSlotsSchema)
   .handler(async ({ input, context }) => {
-    return swapSlots(input, context.user.id);
+    const slots = await MatchDAL.swapSlots(input, context.user.id);
+    return slots;
   });
