+++ src/orpc/tournament-athletes/index.ts
@@ -3,21 +3,16 @@ import {
   BulkAddAthletesSchema,
   ListTournamentAthletesSchema,
   UpdateTournamentAthleteSchema,
-} from './tournament-athletes.dto';
-import {
-  bulkCreate,
-  bulkRemoveTournamentAthletes,
-  findByTournamentId,
-  removeTournamentAthlete,
-  updateTournamentAthlete,
-} from './tournament-athletes.dal';
+} from './dto';
+import { TournamentAthleteDAL } from './dal';
 import { authedProcedure } from '@/orpc/middleware';
 import { prisma } from '@/lib/db';
 
 export const listTournamentAthletes = authedProcedure
   .input(ListTournamentAthletesSchema)
   .handler(async ({ input }) => {
-    return findByTournamentId(input);
+    const athletes = await TournamentAthleteDAL.findByTournamentId(input);
+    return athletes;
   });
 
 export const bulkAddAthletes = authedProcedure
@@ -29,10 +24,12 @@ export const bulkAddAthletes = authedProcedure
       where: { id: { in: athleteProfileIds } },
     });
 
-    const created = await bulkCreate(tournamentId, profiles);
+    const created = await TournamentAthleteDAL.bulkCreate(
+      tournamentId,
+      profiles
+    );
     const added = created.length;
 
-    // autoAssign stub: group constraints not implemented yet (Groups tab is a future item)
     return { added, assigned: 0, unassigned: added };
   });
 
@@ -40,17 +37,27 @@ export const updateTournamentAthleteRecord = authedProcedure
   .input(UpdateTournamentAthleteSchema)
   .handler(async ({ input }) => {
     const { id, ...data } = input;
-    return updateTournamentAthlete(id, data);
+    const athlete = await TournamentAthleteDAL.updateTournamentAthlete(
+      id,
+      data
+    );
+    return athlete;
   });
 
 export const removeTournamentAthleteRecord = authedProcedure
   .input(z.object({ id: z.string() }))
   .handler(async ({ input }) => {
-    return removeTournamentAthlete(input.id);
+    const athlete = await TournamentAthleteDAL.removeTournamentAthlete(
+      input.id
+    );
+    return athlete;
   });
 
 export const bulkRemoveTournamentAthleteRecords = authedProcedure
   .input(z.object({ ids: z.array(z.string()).min(1) }))
   .handler(async ({ input }) => {
-    return bulkRemoveTournamentAthletes(input.ids);
+    const count = await TournamentAthleteDAL.bulkRemoveTournamentAthletes(
+      input.ids
+    );
+    return count;
   });
