+++ src/orpc/tournaments/index.ts
@@ -4,27 +4,20 @@ import {
   ListTournamentsSchema,
   SetTournamentStatusSchema,
   UpdateTournamentSchema,
-} from './tournaments.dto';
-import {
-  create,
-  deleteTournament,
-  findById,
-  findMany,
-  setStatus,
-  update,
-} from './tournaments.dal';
+} from './dto';
+import { TournamentDAL } from './dal';
 import { authedProcedure } from '@/orpc/middleware';
 
 export const listTournaments = authedProcedure
   .input(ListTournamentsSchema)
   .handler(async ({ input }) => {
-    return findMany(input);
+    return TournamentDAL.findMany(input);
   });
 
 export const getTournament = authedProcedure
   .input(z.object({ id: z.string() }))
   .handler(async ({ input }) => {
-    const tournament = await findById(input.id);
+    const tournament = await TournamentDAL.findById(input.id);
     if (!tournament) {
       throw new Error('Tournament not found');
     }
@@ -34,27 +27,31 @@ export const getTournament = authedProcedure
 export const createTournament = authedProcedure
   .input(CreateTournamentSchema)
   .handler(async ({ input }) => {
-    return create(input);
+    const tournament = await TournamentDAL.create(input);
+    return tournament;
   });
 
 export const updateTournament = authedProcedure
   .input(UpdateTournamentSchema)
   .handler(async ({ input }) => {
     const { id, ...data } = input;
-    return update(id, data);
+    const tournament = await TournamentDAL.update(id, data);
+    return tournament;
   });
 
 export const setTournamentStatus = authedProcedure
   .input(SetTournamentStatusSchema)
   .handler(async ({ input, context }) => {
-    return setStatus({
+    const tournament = await TournamentDAL.setStatus({
       ...input,
       adminId: context.user.id,
     });
+    return tournament;
   });
 
 export const removeTournament = authedProcedure
   .input(z.object({ id: z.string() }))
   .handler(async ({ input }) => {
-    return deleteTournament(input.id);
+    const tournament = await TournamentDAL.deleteTournament(input.id);
+    return tournament;
   });
